refactor(backend): isolate code studio delivery prompt
- Move the Code Studio delivery-first answer contract into a focused prompt helper
- Preserve the direct_prompts compatibility import surface for graph/runtime bindings
- Update the runtime cleanup audit for issue #545
